pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tlcm-engine"
version = "0.5.1"
description = "Temporal Layered Context Memory Engine - Biological Edge-Optimized memory for AI Agents"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Collins Somtochukwu", email = "collins@example.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "chromadb>=0.5.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "ollama>=0.2.0",
    "google-genai>=1.70.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.0.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "psycopg2-binary>=2.9.9"
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "matplotlib>=3.7.0"
]

[project.scripts]
tlcm = "tlcm:app"
tlcm-server = "server.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "server*", "integrations*", "tlcm_client*"]
